Refactor to use a single loop when calculating data.
